refactor(pr): PR 列表/选中/审批/合并/刷新归 features/pr 的 usePullRequests
App 主入口承载了大量 PR 领域业务（prs 列表 + 选中、reloadPrs / triggerRefresh、
审批决断 setSelectedPrStatus、合并 mergeSelectedPr、focus 刷新），与「组合根」职责
混淆。抽到 features/pr/hooks/usePullRequests：领域内聚地管列表生命周期 + 选中态 +
详情动作；App 仅注入连接摘要与就绪态，bootstrap / 向导完成时经 setPrs 注入初始列表、
poll tick 时调 reloadPrs。无逻辑/界面变更。
